实现服务端session失效之后，发起重新登陆

+++ app/src/main/java/indi/nonoas/xbh/greendao/AccBalanceDao.java
@@ -26,7 +26,7 @@ public class AccBalanceDao extends AbstractDao<AccBalance, Long> {
     public static class Properties {
         public final static Property SerialNo = new Property(0, Long.class, "serialNo", true, "_id");
         public final static Property AccNo = new Property(1, String.class, "accNo", false, "ACC_NO");
-        public final static Property Date = new Property(2, Long.class, "date", false, "DATE");
+        public final static Property Date = new Property(2, Integer.class, "date", false, "DATE");
         public final static Property UserId = new Property(3, String.class, "userId", false, "USER_ID");
         public final static Property AccName = new Property(4, String.class, "accName", false, "ACC_NAME");
         public final static Property Balance = new Property(5, String.class, "balance", false, "BALANCE");
@@ -76,7 +76,7 @@ public class AccBalanceDao extends AbstractDao<AccBalance, Long> {
             stmt.bindString(2, accNo);
         }
  
-        Long date = entity.getDate();
+        Integer date = entity.getDate();
         if (date != null) {
             stmt.bindLong(3, date);
         }
@@ -111,7 +111,7 @@ public class AccBalanceDao extends AbstractDao<AccBalance, Long> {
             stmt.bindString(2, accNo);
         }
  
-        Long date = entity.getDate();
+        Integer date = entity.getDate();
         if (date != null) {
             stmt.bindLong(3, date);
         }
@@ -142,7 +142,7 @@ public class AccBalanceDao extends AbstractDao<AccBalance, Long> {
         AccBalance entity = new AccBalance( //
             cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0), // serialNo
             cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1), // accNo
-            cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2), // date
+            cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2), // date
             cursor.isNull(offset + 3) ? null : cursor.getString(offset + 3), // userId
             cursor.isNull(offset + 4) ? null : cursor.getString(offset + 4), // accName
             cursor.isNull(offset + 5) ? null : cursor.getString(offset + 5) // balance
@@ -154,7 +154,7 @@ public class AccBalanceDao extends AbstractDao<AccBalance, Long> {
     public void readEntity(Cursor cursor, AccBalance entity, int offset) {
         entity.setSerialNo(cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0));
         entity.setAccNo(cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1));
-        entity.setDate(cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2));
+        entity.setDate(cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2));
         entity.setUserId(cursor.isNull(offset + 3) ? null : cursor.getString(offset + 3));
         entity.setAccName(cursor.isNull(offset + 4) ? null : cursor.getString(offset + 4));
         entity.setBalance(cursor.isNull(offset + 5) ? null : cursor.getString(offset + 5));
